# Switcher cache (improves performance)