--- a/wsmatrix@martin.zurowietz.de/prefs.js
+++ b/wsmatrix@martin.zurowietz.de/prefs.js
@@ -46,6 +46,7 @@ var PrefsWidget = new GObject.Class({
       return [
         'show-thumbnails',
         'show-workspace-names',
+        'cache-switcher',
       ];
    },