I think this is a very interesting PR, it provides an option in settings to enable switcher caching. The GNOME team seems to be using something similar to improve `activities view` performance. I am still not 100% sure whether this will have any side effects or not, I have been playing with it for around and it seems to play well with the `Super+Tab` shortcut that displays the `AppSwitcher`, there is a code that prevents them from being shown up together, and so a one must be destroyed before displaying the other popup.

The motivation behind this PR is only the `show thumbnails` option, thumbnails seem to consume a lot of processing power to be generated, and current implementation **rerenders thumbnails** serveral times while displaying the `workspace switcher`.

Caching will only rerender the `workspace switcher`  if some settings changed or an expected update was detected, like the `style-changed` event. This will make first display of the switcher much slower than the next cached switcher display.

Here is a performance check
`note: the more the workspaces are, the longer time it takes the extension to create them, in a case of 5x5 grid the performance will be much worse. This test was conducted on a 3x3 grid `

```
with caching disabled and thumbnails enabled:
gnome-shell[40250]: time taken to display switcher 170 ms
gnome-shell[40250]: time taken to display switcher 148 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 153 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 135 ms
gnome-shell[40250]: time taken to display switcher 145 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 152 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 132 ms
gnome-shell[40250]: time taken to display switcher 146 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 152 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 135 ms
gnome-shell[40250]: time taken to display switcher 146 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 153 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 135 ms
gnome-shell[40250]: time taken to display switcher 147 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 156 ms
gnome-shell[40250]: complete destroy

=========================================================

with caching enabled and thumbnails enabled:
gnome-shell[40250]: time taken to display switcher 140 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 159 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy

=========================================================
=========================================================
=========================================================

with caching disabled and thumbnails disabled:
gnome-shell[40250]: time taken to display switcher 2 ms
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 5 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 7 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 2 ms
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 2 ms
gnome-shell[40250]: time taken to display switcher 2 ms
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: time taken to display switcher 3 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: complete destroy

=========================================================

with caching enabled and thumbnails disabled:
gnome-shell[40250]: rerenderign switcher ...
gnome-shell[40250]: complete destroy
gnome-shell[40250]: time taken to display switcher 4 ms
gnome-shell[40250]: time taken to display switcher 5 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: style changed to 8
gnome-shell[40250]: time taken to display switcher 8 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 1 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: time taken to display switcher 0 ms
gnome-shell[40250]: simple destroy

```

This PR needs more testing. Also, this feature is disabled by default, and must be enabled in preferences.