@mzur 2019-05-04
Interesting. The delay is not as high on my machine but the improvement with the caching is still obvious:

```
gnome-shell[1088]: time taken to display switcher 48 ms
gnome-shell[1088]: time taken to display switcher 59 ms
gnome-shell[1088]: style changed to 8
gnome-shell[1088]: time taken to display switcher 67 ms
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: time taken to display switcher 61 ms
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: time taken to display switcher 65 ms
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: time taken to display switcher 60 ms
gnome-shell[1088]: time taken to display switcher 62 ms
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: time taken to display switcher 65 ms
gnome-shell[1088]: time taken to display switcher 60 ms
gnome-shell[1088]: time taken to display switcher 61 ms
gnome-shell[1088]: time taken to display switcher 60 ms
gnome-shell[1088]: time taken to display switcher 62 ms
gnome-shell[1088]: complete destroy

-- enable caching, restart gnome --

gnome-shell[1088]: time taken to display switcher 51 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: style changed to 8
gnome-shell[1088]: time taken to display switcher 63 ms
gnome-shell[1088]: rerenderign switcher ...
gnome-shell[1088]: complete destroy
gnome-shell[1088]: time taken to display switcher 48 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: style changed to 8
gnome-shell[1088]: time taken to display switcher 58 ms
gnome-shell[1088]: time taken to display switcher 1 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 1 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 1 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 1 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 0 ms
gnome-shell[1088]: time taken to display switcher 1 ms
gnome-shell[1088]: simple destroy
```

I'll have a closer look at your code later.