@ebeem 2019-05-04
I noticed having more windows open and more workspaces count, makes rendering the switcher slower.
So if you have a 5x5 grid, it will take much longer to render the thumbnails.

My case above is 3x3 grid with many windows open.