@mzur 2019-05-05
Okay, here are my comments:

1. I don't like this as an optional feature. Either it works well and is always enabled or it doesn't work well and is not implemented at all.
2. I believe the switcher without thumbnails isn't worth bothering. The difference between 5 ms and 1 ms is not noticeable, I'd say.
3. I think the switcher should still be destroyed properly so it doesn't take up any resources when it is not needed. In my opinion, the initial delay when the switcher is opened is less critical than the delay when workspaces are switched.

Based on your code and these comments, I implemented a simplified mechanism for the caching in 
c4c0147. There is no caching for the switcher without thumbnails and the switcher takes a little longer to render initially. What do you think of that?

Here are the timings for a 5x5 grid:
```
-- without caching --
gnome-shell[1106]: time taken to display switcher 96 ms (initialize)
gnome-shell[1106]: time taken to display switcher 110 ms
gnome-shell[1106]: time taken to display switcher 127 ms
gnome-shell[1106]: time taken to display switcher 126 ms
gnome-shell[1106]: time taken to display switcher 124 ms
gnome-shell[1106]: time taken to display switcher 130 ms
gnome-shell[1106]: time taken to display switcher 103 ms
gnome-shell[1106]: time taken to display switcher 113 ms
gnome-shell[1106]: time taken to display switcher 137 ms
gnome-shell[1106]: time taken to display switcher 129 ms
gnome-shell[1106]: time taken to display switcher 129 ms
gnome-shell[1106]: time taken to display switcher 130 ms
gnome-shell[1106]: time taken to display switcher 92 ms (initialize)
gnome-shell[1106]: time taken to display switcher 111 ms
gnome-shell[1106]: time taken to display switcher 126 ms
gnome-shell[1106]: time taken to display switcher 125 ms
gnome-shell[1106]: time taken to display switcher 126 ms
gnome-shell[1106]: time taken to display switcher 131 ms

-- with caching --
gnome-shell[1106]: time taken to display switcher 98 ms (initialize)
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 0 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 96 ms (initialize)
gnome-shell[1106]: time taken to display switcher 0 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 0 ms
gnome-shell[1106]: time taken to display switcher 0 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 1 ms
gnome-shell[1106]: time taken to display switcher 0 ms
gnome-shell[1106]: time taken to display switcher 0 ms

-- without thumbnails and caching --
gnome-shell[1106]: time taken to display switcher 6 ms (initialize)
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 2 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 4 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 6 ms (initialize)
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 2 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 4 ms
gnome-shell[1106]: time taken to display switcher 3 ms
gnome-shell[1106]: time taken to display switcher 3 m
```