@ebeem 2019-05-05
regarding point (1), I wanted to make this always enabled, but I couldn't completely make sure that this feature is stable and working with no bugs. The main reason why I am not confident about stability is because of point(3), I am caching the switcher and not destroying it.  I actually only did the `_redisplay` part first, which made a noticeable difference in performance for me. but after caching the switcher as well, it became very smooth, and it looked perfect to me.

I am pretty sure it works well if we discard the switcher caching part and keep the `_redisplay` only, but the optional experimental feature is there so others can test it and report bugs before we confidently enable it by default.

and point (2) is totally valid. I just wanted to have a solution that works for both, but it doesn't really have a great impact on performance.

so my opinion is, we can for now test the `_redisplay` part more and merge it with the master branch (without an option to enable/disable it), but it will be perfect if we test the switcher caching part more and try to merge it in the future. 
And as you said, the indicator part doesn't really need caching.